feature(decomp): match ten TUs in wave 2
MapObjFloat, M3UModel, MapCollisionPlane, AnimalManager,
CameraMarioData, CameraCodeControl, CameraTalk, MarioBlend, MActorAnm
and enemytable now fully match and link. UNUSED reconstructions for
MActorAnm changeMtxCalcType, the enemytable TVector_pointer dtor
(TNameRefPtrAryT now instantiates TVector_pointer<T*> per map mangling)
and TObjManager initObjArray. MapEventSink improved to frame-only
residues with the startSoundActor receiver-spelling arbitration
documented in MSound.hpp.

+++ include/Map/MapCollisionPlane.hpp
@@ -13,7 +13,10 @@ public:
 	TBGCheckData* getCheckData(int x, int z, int which) const;
 
 	f32 gridToWorld(int v) { return v * mScale - mExtent; }
-	f32 worldToGrid(f32 v) { return mOneOverScale * (v + mExtent); }
+	// TODO: the truncating int return is required to match checkPlaneGround;
+	// TMapObjPlane::depress needs the fractional part, so it probably
+	// spelled the expression out instead of calling this
+	int worldToGrid(f32 v) { return mOneOverScale * (v + mExtent); }
 
 public:
 	/* 0x0 */ int mGridWidth;
